+++ ConCurrency/FromForeignToPLN.hpp
@@ -6,5 +6,5 @@ class FromForeignToPLN : public ExchangeCurrency
 {
 public:
 	FromForeignToPLN(std::shared_ptr<CashRegister> cashRegister);
-	std::pair<int, double> currencyExchange(int amount, std::string currencyCode) override;
+	std::pair<double, double> currencyExchange(double amount, std::string currencyCode) override;
 };
